+++ packages/jest-expo/CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### 🐛 Bug fixes
 
-- Fix babel config resolution to distinguish project-wide configs (`babel.config.*`) from directory-scoped configs (`.babelrc`). ([#45683](https://github.com/expo/expo/pull/45683) by [@mvincentong](https://github.com/mvincentong))
+- Align babel config resolution with Expo CLI's project-root config lookup. ([#45683](https://github.com/expo/expo/pull/45683) by [@mvincentong](https://github.com/mvincentong))
 
 ### 💡 Others